Moved mappers into separate package.
